+++ Sandswept/Utils/AttackUtils/BasicLaserBeam.cs
@@ -15,18 +15,21 @@ namespace Sandswept.Utils
         private float stopwatch = 0f;
         private float growthStopwatch = 0f;
         private float delay;
-        private BasicLaserInfo info;
+        public BasicLaserInfo info;
         private Transform origin;
         private Transform end;
         private Vector3 targetEndpoint;
+        private Vector3 lastEndpoint;
         private float origWidth = 0f;
+        private float stopwatch2 = 0f;
+        private List<CharacterBody> alreadyDamaged = new();
 
         public BasicLaserBeam(CharacterBody owner, Transform muzzle, BasicLaserInfo info)
         {
             this.info = info;
             delay = 1f / info.TickRate;
             TargetMuzzle = muzzle;
-            DamageCoefficient = info.DamageCoefficient * delay;
+            DamageCoefficient = info.SingleHit ? info.DamageCoefficient : info.DamageCoefficient * delay;
             effectInstance = GameObject.Instantiate(info.EffectPrefab, muzzle.transform.position, Quaternion.identity);
             growthStopwatch = info.ChargeDelay;
             origin = info.OriginIsBase ? effectInstance.transform : effectInstance.GetComponent<ChildLocator>().FindChild(info.OriginName);
@@ -36,6 +39,7 @@ namespace Sandswept.Utils
             Owner = owner;
 
             targetEndpoint = GetEndpoint(out _);
+            lastEndpoint = targetEndpoint;
             end.transform.position = targetEndpoint;
             origin.transform.position = TargetMuzzle.transform.position;
         }
@@ -54,8 +58,20 @@ namespace Sandswept.Utils
 
         public void UpdateVisual(float deltaTime)
         {
+            stopwatch2 += deltaTime;
+
             origin.transform.position = TargetMuzzle.transform.position;
-            end.transform.position = Vector3.MoveTowards(end.transform.position, targetEndpoint, 250f * deltaTime);
+            end.transform.position = Vector3.Lerp(lastEndpoint, targetEndpoint, stopwatch2 / delay);
+
+            if (stopwatch2 >= delay) {
+                stopwatch2 = 0f;
+            }
+
+            if (!firing)
+            {
+                growthStopwatch -= deltaTime;
+                lr.widthMultiplier = Mathf.Max(0f, (growthStopwatch / info.ChargeDelay));
+            }
         }
 
         public void Update(float deltaTime)
@@ -64,6 +80,7 @@ namespace Sandswept.Utils
 
             if (stopwatch >= delay)
             {
+                lastEndpoint = targetEndpoint;
                 targetEndpoint = GetEndpoint(out Vector3 impact);
                 stopwatch = 0f;
 
@@ -79,19 +96,15 @@ namespace Sandswept.Utils
                             scale = 1f
                         }, false);
                     }
-                }
-            }
 
-            if (!firing)
-            {
-                growthStopwatch -= deltaTime;
-                lr.widthMultiplier = Mathf.Max(0f, (growthStopwatch / info.ChargeDelay));
+                    info.ImpactCallback?.Invoke(impact);
+                }
             }
         }
 
         public Vector3 GetEndpoint(out Vector3 unmodified)
         {
-            Vector3 dir = (info.FiringMode == LaserFiringMode.TrackAim) ? Owner.inputBank.aimDirection : TargetMuzzle.forward;
+            Vector3 dir = (info.FiringMode == LaserFiringMode.TrackAim) ? Owner.inputBank.aimDirection : info.UseUP ? TargetMuzzle.up : TargetMuzzle.forward;
             Vector3 pos = (info.FiringMode == LaserFiringMode.TrackAim) ? Owner.inputBank.aimOrigin : TargetMuzzle.position;
             Vector3 endpoint = new Ray(pos, dir).GetPoint(info.MaxRange);
 
@@ -118,6 +131,19 @@ namespace Sandswept.Utils
             attack.falloffModel = BulletAttack.FalloffModel.None;
             attack.isCrit = Util.CheckRoll(Owner.crit, Owner.master);
             attack.stopperMask = LayerIndex.world.mask;
+            attack.hitCallback = (BulletAttack attack, ref BulletAttack.BulletHit hit) => {
+                bool res = BulletAttack.defaultHitCallback(attack, ref hit);
+
+                if (info.SingleHit && hit.hurtBox && hit.hurtBox.healthComponent) {
+                    if (alreadyDamaged.Contains(hit.hurtBox.healthComponent.body)) {
+                        return false;
+                    }
+
+                    alreadyDamaged.Add(hit.hitHurtBox.healthComponent.body);
+                }
+
+                return res;
+            };
 
             return attack;
         }
@@ -142,6 +168,9 @@ namespace Sandswept.Utils
         public LaserFiringMode FiringMode = LaserFiringMode.Straight;
         public float MaxRange = 60f;
         public GameObject ImpactEffect;
+        public bool SingleHit = false;
+        public bool UseUP = false;
+        public Action<Vector3> ImpactCallback;
     }
 
     public enum LaserFiringMode